requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
sqlalchemy>=2.0
aiosqlite>=0.20
httpx>=0.27
openai>=1.12
PyJWT>=2.8
slowapi>=0.1.9
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
